fix a find-all-replace typo in Nevig's config
When i fixed the filepaths in the Kialla system i changed every "\" to a "/" to hopefully fix Linux compatibility. But I also changed all "/n" to "\n" (newlines in planet descriptions only work with a backslash). Unfortunately this had also put a backslash and lower case "n" in all of Nevig's filepaths. The filepaths should now be fixed.